Update HelloAppDelegate.m
Fixing the memory leak. Fixing the style. Fixing an incorrect singleton paradigm that would lead to a crash if the application delegate ever changes.

Also ViewController should be setting itself as rdio.delegate and rdio.player.delegate. It is inappropriate to assign an instance as a delegate when you do not control the lifecycle of the instance.